requirements2.txt
tqdm==4.61.0
albumentations==1.0.0
fvcore==0.1.5.post20210518
hydra-core==1.1.0rc1
jpeg4py==0.1.4
kornia==0.5.3
matplotlib==3.4.2
numpy==1.20.3
omegaconf==2.1.0.rc1
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
scipy==1.6.3
seaborn==0.11.1
torchsummary==1.5.1
wandb==0.10.31
